% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{acc_from_obd}
\alias{acc_from_obd}
\title{Calculate acceleration from OBD2 data}
\usage{
acc_from_obd(obd_speed)
}
\arguments{
\item{obd_speed}{obd_speed which is splined already}
}
\value{
a estimated longitudinal acceleration from OBD2 speed data
}
\description{
Calculate acceleration from OBD2 data
}
