% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{adjust_alt}
\alias{adjust_alt}
\title{altitude data length adjustment.}
\usage{
adjust_alt(time_data, alt_data)
}
\arguments{
\item{time_data}{a data which has time column}

\item{alt_data}{an altidute data}
}
\value{
length adjusted alt data
}
\description{
altitude data length adjustment.
}
