% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_telematics}
\alias{calibrate_telematics}
\title{Calibarate telematics data}
\usage{
calibrate_telematics(file_path)
}
\arguments{
\item{file_path}{file path of telematics data}
}
\value{
calibrated GPS data and accelerometer x-axis and y-axis data
}
\description{
Calibarate telematics data
}
