% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\name{convert_time}
\alias{convert_time}
\title{convert time stamp into seconds}
\usage{
convert_time(timelist)
}
\arguments{
\item{timelist}{time value in list consists of 3 elements;
the first element is hour, the second element is min, the last one is seconds.}
}
\value{
seconds corresponds with the input hour
}
\description{
convert time stamp into seconds
}
