% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{converTOmeter}
\alias{converTOmeter}
\alias{convertTOmeter}
\title{Calculate distance between two GPS points into meter scale}
\usage{
convertTOmeter(pointVector)
}
\arguments{
\item{pointVector}{two GPS coordinate points}
}
\value{
distance between two GPS points
}
\description{
Calculate distance between two GPS points into meter scale
}
