% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Cat citation message when the package is loading}
\usage{
.onAttach(libname, pkgname)
}
\value{
start up message
}
\description{
Cat citation message when the package is loading
}
