% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_map}
\alias{draw_map}
\title{draw GPS points}
\usage{
draw_map(map_obj, gps_data)
}
\arguments{
\item{map_obj}{the object from google map}

\item{gps_data}{GPS data}
}
\value{
ggmap object
}
\description{
draw GPS points
}
