% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{draw_vaHeatmap}
\alias{draw_vaHeatmap}
\title{Draw a v-a heatmap for telematics data}
\usage{
draw_vaHeatmap(telematics_data, drawRange = c(0, 80))
}
\arguments{
\item{telematics_data}{telematics data with a_lon, a_lat coloums}

\item{drawRange}{vector for drawing range of heat map}
}
\value{
telematics heatmap drawn by ggplot
}
\description{
Draw a v-a heatmap for telematics data
}
