% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{drawHeatmap}
\alias{drawHeatmap}
\title{Draw a heatmap for telematics data}
\usage{
drawHeatmap(telematics_data, drawRange = c(4, 4))
}
\arguments{
\item{telematics_data}{telematics data with a_lon, a_lat coloums}

\item{drawRange}{vector for drawing range of heat map}
}
\value{
telematics heatmap drawn by ggplot
}
\description{
Draw a heatmap for telematics data
}
