% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{get_NE}
\alias{get_NE}
\title{Convert accelerometer value into lat., long. acceleration}
\usage{
get_NE(acc_data, heading_data)
}
\arguments{
\item{acc_data}{accelerometer data which has time, x and y}

\item{heading_data}{heading data based on magnetometer which has time and trueheading}
}
\value{
a dataframe which has latudinal and longitudinal acceleration data.
}
\description{
Convert accelerometer value into lat., long. acceleration
}
