% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_obd.R
\name{get_obd_trip}
\alias{get_obd_trip}
\title{Sync OBD speed data with GPS speed data}
\usage{
get_obd_trip(file_path)
}
\arguments{
\item{file_path}{file path of telematics data}
}
\value{
synced OBD speed with time and speed of the vehicle
}
\description{
Sync OBD speed data with GPS speed data
}
