% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{get_speed}
\alias{get_speed}
\title{Calculate speed(mph) from GPS coordinates}
\usage{
get_speed(gps_coordinates)
}
\arguments{
\item{gps_coordinates}{GPS coordinates}
}
\value{
speed recorded as mile per hour
}
\description{
Calculate speed(mph) from GPS coordinates
}
