% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_smoothing.R
\name{get_stopvec}
\alias{get_stopvec}
\title{Making stop moments index from stop information}
\usage{
get_stopvec(stop_info)
}
\arguments{
\item{stop_info}{stop information}
}
\value{
an index stop vector
}
\description{
Making stop moments index from stop information
}
