% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\name{get_trip}
\alias{get_trip}
\title{Loading the telematics trip csv file.}
\usage{
get_trip(data_path, data_option = 0, sync_time = NULL)
}
\arguments{
\item{data_path}{a path that the trip csv file located}

\item{data_option}{the variables that the user want to select from the data. \cr
The following list explains the options; \cr
Option 0 : all \cr
Option 1 : gps \cr
Option 2 : accel xyz \cr
Option 3 : linacc xyz \cr
Option 4 : gyro xyz \cr
Option 5 : roll, pitch, yaw \cr
Option 6 : speed, heading
Option 7 : magnetometer x, y, z \cr}

\item{sync_time}{this arguments will be used when the recorded data has different time line with video.}
}
\description{
get_trip function will read a csv file which records the driving trip data
}
