% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{get_tripmap}
\alias{get_tripmap}
\title{get trip map from google map}
\usage{
get_tripmap(start_t, end_t, gps_data, api)
}
\arguments{
\item{start_t}{starting time}

\item{end_t}{end time}

\item{gps_data}{GPS data}

\item{api}{google map api}
}
\value{
ggmap object
}
\description{
get trip map from google map
}
