% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter}
\alias{kalmanfilter}
\title{Kalmanfilter general}
\usage{
kalmanfilter(x, P, z, u, A, B, H, Q, R)
}
\arguments{
\item{x}{observation from sensor 1}

\item{P}{variance of observation from sensor 1 at the previous time step}

\item{z}{observation from sensor 2}

\item{u}{control vector}

\item{A}{scale constant related to x}

\item{B}{scale constant related to u}

\item{H}{scale constant between sensor 1 and 2}

\item{Q}{noise of the observation from sensor 1}

\item{R}{noise of the observation from sensor 2}
}
\value{
the list of two value; optimal estimate for x and P
}
\description{
Kalmanfilter general
}
