% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter_1D}
\alias{kalmanfilter_1D}
\title{1-D Kalmanfilter}
\usage{
kalmanfilter_1D(x, u, a, b, z, h, P, Q, R)
}
\arguments{
\item{x}{observation from sensor 1}

\item{u}{control vector}

\item{a}{scale constant related to x}

\item{b}{scale constant related to u}

\item{z}{observation from sensor 2}

\item{h}{scale constant between sensor 1 and 2}

\item{P}{variance of observation from sensor 1 at the previous time step}

\item{Q}{noise of the observation from sensor 1}

\item{R}{noise of the observation from sensor 2}
}
\value{
the list of two value; optimal estimate for x and P
}
\description{
1-D Kalmanfilter
}
