% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter_ax}
\alias{kalmanfilter_ax}
\title{Calibarate accelerometer x-axis using Gyroscope data}
\usage{
kalmanfilter_ax(acc_data, gyro_data, smth_speed_data)
}
\arguments{
\item{acc_data}{accelerometer data containing time, x, y, z}

\item{gyro_data}{Gyroscope data containing time, x, y, z}

\item{smth_speed_data}{speed data which has same number of points with accelerometer data.}
}
\value{
calibrated x-axis accelerometer data
}
\description{
Calibarate accelerometer x-axis using Gyroscope data
}
