% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter_gps}
\alias{kalmanfilter_gps}
\title{Calibarate GPS data using accelerometer data}
\usage{
kalmanfilter_gps(gps_data, acc_data, heading_data)
}
\arguments{
\item{gps_data}{GPS data which has the same number of points with accelerometer data}

\item{acc_data}{accelerometer data containing time, x, y, z}

\item{heading_data}{heading data based on magnetometer which has time and trueheading data.}
}
\value{
calibrated GPS data
}
\description{
Calibarate GPS data using accelerometer data
}
