% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter_roadgrade}
\alias{kalmanfilter_roadgrade}
\title{kalman filter with road grade for lon. accelerometer}
\usage{
kalmanfilter_roadgrade(angle_data)
}
\arguments{
\item{angle_data}{angle data from gyro}
}
\value{
filtered road grade
}
\description{
kalman filter with road grade for lon. accelerometer
}
