% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{kalmanfilter_withalpha}
\alias{kalmanfilter_withalpha}
\title{kalman filter with road grade for lon. accelerometer}
\usage{
kalmanfilter_withalpha(acc_data, speed_data, angle_data, alt_data)
}
\arguments{
\item{acc_data}{acc data}

\item{speed_data}{speed data from gps}

\item{angle_data}{angle data from gyro}

\item{alt_data}{altitude data from baro}
}
\value{
filtered acc data with road grade
}
\description{
kalman filter with road grade for lon. accelerometer
}
