% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\name{load_telematic_data}
\alias{load_telematic_data}
\title{Load telematics data set}
\usage{
load_telematic_data(file_path, all_in_one = FALSE, ...)
}
\arguments{
\item{file_path}{file path to the telematics route}

\item{all_in_one}{if true, load all telematics data in a list, 
otherwise will give you the dataframe}

\item{...}{you can access to the arguments in get_trip() when all_in_one = FALSE}
}
\value{
seconds corresponds with the input hour
}
\description{
Load telematics data set
}
