% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{low_pass}
\alias{low_pass}
\title{Low pass filter}
\usage{
low_pass(input_vec, alpha)
}
\arguments{
\item{input_vec}{input vector that needs to be filtered}

\item{alpha}{constant parameter for low pass rate}
}
\value{
filtered data
}
\description{
Low pass filter
}
