% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm_model.R
\name{make_y}
\alias{make_y}
\title{make y vectors for calibration}
\usage{
make_y(file_path)
}
\arguments{
\item{file_path}{route file path}
}
\value{
dataframe of observations
}
\description{
make y vectors for calibration
}
