% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{OBD2_data}
\alias{OBD2_data}
\title{Read speed data from OBD2 device}
\usage{
OBD2_data(file_path)
}
\arguments{
\item{file_path}{OBD2 data}
}
\value{
a OBD2 data with speed
}
\description{
Read speed data from OBD2 device
}
