% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensordata_plot.R
\name{plot_acc}
\alias{plot_acc}
\alias{plot_speed}
\title{Visualization of accelerometer sensor data}
\usage{
plot_acc(data, rate = 1, tripname = "a trip", ...)

plot_speed(data, tripname = "a trip", showobd = TRUE, ...)
}
\arguments{
\item{data}{speed information dataframe (obd_speed can be added)}

\item{rate}{low pass filter parameter 0 to 1}

\item{tripname}{name of the trip}

\item{...}{arguments supported by base plot}

\item{showobd}{show obd speed info.}
}
\value{
plot of three axes accelerometer data

plot of speed information
}
\description{
Visualization of accelerometer sensor data

Visualization of gps speed data
}
