% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensordata_plot.R
\name{plot_acc_compare}
\alias{plot_acc_compare}
\title{Visualization of telematics data}
\usage{
plot_acc_compare(
  data,
  rate = 0.2,
  sensor_name = c("sensor 1", "sensor 2"),
  ...
)
}
\arguments{
\item{data}{speed information dataframe (obd_speed can be added)}

\item{rate}{low pass filter parameter (smoothing 0 to 1)}

\item{sensor_name}{names of the sensors}

\item{...}{arguments supported by base plot}
}
\value{
comparison plot of two sensor data
}
\description{
Visualization of telematics data
}
