% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensordata_plot.R
\name{plot_telematics}
\alias{plot_telematics}
\title{Visualization of telematics data}
\usage{
plot_telematics(data, rate = 1, methodname = "the calibration", ...)
}
\arguments{
\item{data}{speed information dataframe (obd_speed can be added)}

\item{rate}{low pass filter parameter (smoothing 0 to 1)}

\item{methodname}{name of the method}

\item{...}{arguments supported by base plot}
}
\value{
plot of calibrated telematics data
}
\description{
Visualization of telematics data
}
