% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_smoothing.R
\name{smoothing_gps}
\alias{smoothing_gps}
\title{Smoothing GPS data using smoothing spline}
\usage{
smoothing_gps(gps_data)
}
\arguments{
\item{gps_data}{the collection of GPS coordinates with time info.}
}
\value{
smoothed GPS data with same number of GPS points.
}
\description{
Smoothing GPS data using smoothing spline
}
