% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{smoothing_speed}
\alias{smoothing_speed}
\title{Smooting speed(mph) from speed data}
\usage{
smoothing_speed(speed_data, stop_info)
}
\arguments{
\item{speed_data}{speed_data}

\item{stop_info}{stop information}
}
\value{
smoothed speed data recorded as mile per hour
}
\description{
Smooting speed(mph) from speed data
}
