% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{speed_from_acc}
\alias{speed_from_acc}
\title{Calculate speed(mph) from accelerometer y values}
\usage{
speed_from_acc(time_vec, acc_vec)
}
\arguments{
\item{time_vec}{time vector}

\item{acc_vec}{accelerometer vector}
}
\value{
a speed vector recorded as mile per hour
}
\description{
Calculate speed(mph) from accelerometer y values
}
