% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{speed_from_gps}
\alias{speed_from_gps}
\title{Calculate speed(mph) from GPS coordinates}
\usage{
speed_from_gps(gps_data)
}
\arguments{
\item{gps_data}{GPS coordinates}
}
\value{
speed recorded as mile per hour
}
\description{
Calculate speed(mph) from GPS coordinates
}
