% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{speed_from_obd}
\alias{speed_from_obd}
\title{Spline obd speed data from OBD2 device}
\usage{
speed_from_obd(obd_data, speed_data)
}
\arguments{
\item{obd_data}{OBD2 data}

\item{speed_data}{the speed data that you want to sync, which consists of time and speed}
}
\value{
a estimated speed using spline from OBD2 speed data
}
\description{
Spline obd speed data from OBD2 device
}
