% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_smoothing.R
\name{stop_correction}
\alias{stop_correction}
\title{Correction of GPS based on stop information}
\usage{
stop_correction(gps_data, stop_info)
}
\arguments{
\item{gps_data}{the GPS data}

\item{stop_info}{the information of stop}
}
\value{
A GPS data which does not have different GPS points when the vehicle is stop
}
\description{
Correction of GPS based on stop information
}
