% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalmanfilter.R
\name{stop_correction_acc}
\alias{stop_correction_acc}
\title{Correct accelerometer values for stop}
\usage{
stop_correction_acc(acc_data, stop_info)
}
\arguments{
\item{acc_data}{accelerometer data containing time, x, y, z}

\item{stop_info}{information of stop}
}
\value{
An accelerometer data whose values are zero at stop moments
}
\description{
Correct accelerometer values for stop
}
