% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_smoothing.R
\name{stop_detection}
\alias{stop_detection}
\title{Detecting stops from accelerometer y-axis}
\usage{
stop_detection(acc_data)
}
\arguments{
\item{acc_data}{the accelerometer data}
}
\value{
A dataframe containing the position of stop starts and ends, and length of stop.
}
\description{
Detecting stops from accelerometer y-axis
}
