% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{telematics_data_fromGPS}
\alias{telematics_data_fromGPS}
\title{Preparing a telematics data from GPS data for heatmap drawing}
\usage{
telematics_data_fromGPS(gps_data, speed_data)
}
\arguments{
\item{gps_data}{GPS data}

\item{speed_data}{Speed data}
}
\value{
telematics data object for drawing heatmap
}
\description{
Preparing a telematics data from GPS data for heatmap drawing
}
