% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{track_vis}
\alias{track_vis}
\title{Trip visualization of telematics data}
\usage{
track_vis(gps_data, timepoint = NA, animate = FALSE, api, zoom = 1)
}
\arguments{
\item{gps_data}{gps data dataframe from load_telematics_data}

\item{timepoint}{specific time point in the trip}

\item{animate}{make TRUE for obtaining trip animation}

\item{api}{google map api}

\item{zoom}{zooming parameter}
}
\value{
ggmap object
}
\description{
Trip visualization of telematics data
}
