% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_speed.R
\name{vspeed_from_baro}
\alias{vspeed_from_baro}
\title{Vertical speed from Barometer data}
\usage{
vspeed_from_baro(alt_data)
}
\arguments{
\item{alt_data}{the altitude data}
}
\value{
altitude data with vertical speed
}
\description{
Vertical speed from Barometer data
}
