% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_telematics}
\alias{calibrate_telematics}
\title{Calibarate telematics data}
\usage{
calibrate_telematics(file_path, sync_time = NULL)
}
\arguments{
\item{file_path}{file path of telematics data}

\item{sync_time}{sync_time}
}
\value{
calibrated long. and lat. acceleration of the vehicle
}
\description{
Calibarate telematics data
}
