% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readcsv.R
\name{get_trip}
\alias{get_trip}
\title{loading the telematics trip csv file.}
\usage{
get_trip(data_folder = "", tripN, data_option = 0, sync_time = NULL)
}
\description{
get_trip function will read the csv file which records the driving trip data
}
